Removed aspect dependency which results in no more coping aspect classes in each jar.

+++ openmole/core/org.openmole.core.workflow.implementation/src/main/java/org/openmole/core/workflow/implementation/execution/batch/BatchExecutionJob.java
@@ -54,7 +54,7 @@ public class BatchExecutionJob<JS extends IBatchJobService> extends ExecutionJob
         Activator.getWorkspace().addToConfigurations(UpdateInterval, "PT2M");
     }
 
-    long updateInterval;
+    Integer updateInterval;
     IUpdatableFuture future;
     IBatchJob batchJob;
     Boolean killed = false;
